refactor(FilesHandler): add FilesHandler class to get compare and dataset files
