+++ src/Ml.Cli.WebApp/Paths/FilesHandler.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using Ml.Cli.FileLoader;
+
+namespace Ml.Cli.WebApp.Paths
+{
+    public static class FilesHandler
+    {
+        public static IEnumerable<string> GetFilesFromPaths(string paths, BasePath basePath, IFileLoader fileLoader)
+        {
+            var jsonExtension = ".json";
+            var pathsArray = paths.Split(Separators.CommaSeparator);
+            var fullyQualifiedPaths =
+                pathsArray.Select(path => Path.IsPathRooted(path) ? path : Path.Combine(basePath.Path, path));
+            
+            return fullyQualifiedPaths
+                .SelectMany(fileLoader.EnumerateFiles)
+                .Where(file => Path.GetExtension(file) == jsonExtension);
+        }
+    }
+}
